Add AdminPanelProvider for app and refactor User model for TwoFactorAuthenticatable
The AdminPanelProvider.php was added to provide a more customisable admin panel interface. This is especially handy for implementing more focused authorisation and navigation parameters. Furthermore, the User model was refactored to support two-factor authentication, enhancing the security factor of the application. Several new files were added related to JenisBantuanResource and JenisPelayananResource which will be used to manage these specific resources. KriteriaPelayanan - a new criterion service resource was also created. The server configuration in vite.config.js was updated to support the detection of the server configuration based on the local host. At last, the composer.lock file was updated due to the changes in the packages used.