Add migration files and classes related to 'bantuan_ppks' and 'address'
This commit includes new migration files related to 'bantuan_ppks' and 'address' table.
The 'bantuan_ppks' table is added with related attributes and relationships to the database migration.
Similarly, an 'address' migration file is created with various attributes like 'alamat', 'provinsi', etc.
Models for the respective tables are added with their relationships.

Also, used new MoneyCast class to handle the conversion of integers to floating-point money values.
Moreover, other changes include the addition of a few traits, enums, and additional attributes to existing models for functionality needs.

The changes are essential for implementing the new feature related to Bantuan PPKS (Pusat Pemberdayaan Keluarga dan Sosial) providing aid to households, where addresses of beneficiaries need to be handled. As a part of this, the 'jenis_bantuan', 'jenis_pelayanan', etc., are managed. The Keluarga model now accepts an addressable morph relationship. This helps in flexible manipulation of households' addresses.