Add database migration file address 2 deletion-be and refactor echo.js file
This commit removes the column 'alamat2' from the 'alamat' table as it was deemed redundant. This change was reflected within the Laravel migration file.

The commit also includes a significant reduction of the echo.js file size. The original file was causing performance issues due to the large size. Code repetitions have been removed and replaced with variables and loops for better performance.