"Removed unused black logo svg file"
The file 'public/images/logo/logo-black.svg' has been removed. The reasons for this change are to keep the codebase clean and increase maintainability, as the file is unused and it takes up space unnecessarily in the repository. The deletion does not affect the functionality or appearance of the application.