Remove Bantuan PKH/BPNT codebase, models, migrations, policies, resources, pages
This commit removes the database migrations, models, and policy from 'bantuan_pkh_bpnt' (PHP Social Assistance). The action was needed due to a restructuring of the database, where the separation for 'bantuan_pkh_bpnt' into different tables ('bantuan_pkh' and 'bantuan_bpnt') was decided. The reason for such changes is to provide a more coherent database structure that would increase the system's efficiency. In addition, the Filament resource files, associated with BantuanPkhBpnt, were modified to match the newly restructured schema and models. These modifications will facilitate the management of the 'bantuan_pkh' and 'bantuan_bpnt' tables.