Update PenggantiRastra resource schema and table
Expanded the resource schema and table for "PenggantiRastra" in the Filament admin panel. Required fields have been added to the schema form and searchable/sortable columns defined for the table view. A few imports were also added to access necessary classes and components. The relationship with "Keluarga" model was also modified and the AlasanEnum was used for select options on the form. The change was needed to ensure detailed data entry and display on PenggantiRastra records for better administrative management. The filename of page "ManagePenggantiRastras" was changed to "ManagePenggantiRastra" to maintain consistency with the class name. The unused "