Remove unused PHP class and create new classes for DataPesertaJamkesda management.
This commit deletes the PHP class for `ManageBantuanPkh` as it is unused in the system. New classes to manage `DataPesertaJamkesda` are created, including functionality for uploading and processing data. A new chart is also added to visualize the data. The update includes a thorough code refactor and updates to modal labels for improve clarity and user experience. Functionality for filtering actions in `BantuanBpjsOverview` has been added, allowing users to filter by date range.